requirements.txt
pandas
requests
mysql-connector-python
streamlit
python-dotenv